requirements.txt
asgiref==3.8.1
Django==5.0.6
django-ratelimit==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
shortuuid==1.0.13
sqlparse==0.5.0
tzdata==2024.1
django_redis==5.2.0
whitenoise==6.6.0 
gunicorn==21.2.0 
django-cors-headers==4.4.0